checkpoint changes
