+++ service/src/main/java/bio/terra/catalog/service/DatasetService.java
@@ -4,7 +4,6 @@ import bio.terra.catalog.common.StorageSystem;
 import bio.terra.catalog.datarepo.DatarepoService;
 import bio.terra.catalog.iam.SamAction;
 import bio.terra.catalog.iam.SamService;
-import bio.terra.catalog.model.ColumnModel;
 import bio.terra.catalog.model.DatasetPreviewTable;
 import bio.terra.catalog.model.DatasetPreviewTablesResponse;
 import bio.terra.catalog.model.DatasetsListResponse;
@@ -15,18 +14,13 @@ import bio.terra.catalog.service.dataset.DatasetAccessLevel;
 import bio.terra.catalog.service.dataset.DatasetDao;
 import bio.terra.catalog.service.dataset.DatasetId;
 import bio.terra.common.exception.BadRequestException;
-import bio.terra.common.exception.NotFoundException;
 import bio.terra.common.exception.UnauthorizedException;
 import bio.terra.common.iam.AuthenticatedUserRequest;
-import bio.terra.datarepo.model.TableModel;
-import bio.terra.rawls.model.Entity;
-import bio.terra.rawls.model.EntityTypeMetadata;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 import com.fasterxml.jackson.databind.node.TextNode;
 import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
@@ -36,7 +30,7 @@ import org.springframework.stereotype.Service;
 @Service
 public class DatasetService {
   private final DatarepoService datarepoService;
-  private final RawlsService rawlsService;
+  public final RawlsService rawlsService;
   private final SamService samService;
   private final DatasetDao datasetDao;
   private final ObjectMapper objectMapper;
@@ -57,31 +51,13 @@ public class DatasetService {
     this.objectMapper = objectMapper;
   }
 
-  public static class IllegalMetadataException extends RuntimeException {
-    public IllegalMetadataException(Throwable cause) {
-      super(cause);
-    }
-  }
+  private record DatasetWithAccessLevel(Dataset dataset, DatasetAccessLevel accessLevel) {}
 
-  private class DatasetWithAccessLevel {
-    private final Dataset dataset;
-    private final DatasetAccessLevel accessLevel;
-
-    public DatasetWithAccessLevel(Dataset dataset, DatasetAccessLevel accessLevel) {
-      this.dataset = dataset;
-      this.accessLevel = accessLevel;
-    }
-
-    public Dataset getDataset() {
-      return dataset;
-    }
-
-    public Object convertToObject() {
-      ObjectNode node = toJsonNode(dataset.metadata());
-      node.set("accessLevel", TextNode.valueOf(String.valueOf(accessLevel)));
-      node.set("id", TextNode.valueOf(dataset.id().toValue()));
-      return node;
-    }
+  public Object convertDatasetToObject(DatasetWithAccessLevel dataset) {
+    ObjectNode node = toJsonNode(dataset.dataset.metadata());
+    node.set("accessLevel", TextNode.valueOf(String.valueOf(dataset.accessLevel)));
+    node.set("id", TextNode.valueOf(dataset.dataset.id().toValue()));
+    return node;
   }
 
   private ObjectNode toJsonNode(String json) {
@@ -125,14 +101,14 @@ public class DatasetService {
               .filter(
                   datasetWithAccessLevel ->
                       !datasets.stream()
-                          .map(datasetInList -> datasetInList.getDataset().id())
+                          .map(datasetInList -> datasetInList.dataset.id())
                           .toList()
-                          .contains(datasetWithAccessLevel.getDataset().id()))
+                          .contains(datasetWithAccessLevel.dataset.id()))
               .toList());
     }
     var response = new DatasetsListResponse();
 
-    response.setResult(datasets.stream().map(DatasetWithAccessLevel::convertToObject).toList());
+    response.setResult(datasets.stream().map(this::convertDatasetToObject).toList());
     return response;
   }
 
@@ -152,10 +128,8 @@ public class DatasetService {
   private List<TableMetadata> generateDatasetTables(
       AuthenticatedUserRequest user, Dataset dataset) {
     return switch (dataset.storageSystem()) {
-      case TERRA_DATA_REPO -> convertDatarepoTablesToCatalogTables(
-          datarepoService.getPreviewTables(user, dataset.storageSourceId()).getTables());
-      case TERRA_WORKSPACE -> convertRawlsTablesToCatalogTables(
-          rawlsService.entityMetadata(user, dataset.storageSourceId()));
+      case TERRA_DATA_REPO -> datarepoService.getPreviewTables(user, dataset.storageSourceId());
+      case TERRA_WORKSPACE -> rawlsService.getPreviewTables(user, dataset.storageSourceId());
       case EXTERNAL -> List.of();
     };
   }
@@ -163,64 +137,10 @@ public class DatasetService {
   private DatasetPreviewTable generateDatasetTablePreview(
       AuthenticatedUserRequest user, Dataset dataset, String tableName) {
     return switch (dataset.storageSystem()) {
-      case TERRA_DATA_REPO -> generateDatarepoTable(user, dataset, tableName, MAX_ROWS);
-      case TERRA_WORKSPACE -> generateRawlsTable(user, dataset, tableName, MAX_ROWS);
+      case TERRA_DATA_REPO -> datarepoService.previewTable(user, dataset, tableName, MAX_ROWS);
+      case TERRA_WORKSPACE -> rawlsService.previewTable(user, dataset, tableName, MAX_ROWS);
       case EXTERNAL -> new DatasetPreviewTable();
     };
-  }
-
-  private DatasetPreviewTable generateDatarepoTable(
-      AuthenticatedUserRequest user, Dataset dataset, String tableName, int maxRows) {
-    return new DatasetPreviewTable()
-        .columns(
-            datarepoService.getPreviewTables(user, dataset.storageSourceId()).getTables().stream()
-                .filter(tableModel -> tableModel.getName().equals(tableName))
-                .findFirst()
-                .orElseThrow(
-                    () ->
-                        new NotFoundException(
-                            String.format(
-                                "Table %s is not found for dataset %s", tableName, dataset.id())))
-                .getColumns()
-                .stream()
-                .map(DatasetService::convertDatarepoColumnModelToCatalogColumnModel)
-                .toList())
-        .rows(
-            datarepoService
-                .getPreviewTable(user, dataset.storageSourceId(), tableName, maxRows)
-                .getResult());
-  }
-
-  private DatasetPreviewTable generateRawlsTable(
-      AuthenticatedUserRequest user, Dataset dataset, String tableName, int maxRows) {
-    Map<String, EntityTypeMetadata> entities =
-        rawlsService.entityMetadata(user, dataset.storageSourceId());
-    EntityTypeMetadata tableMetadata = entities.get(tableName);
-    return new DatasetPreviewTable()
-        .columns(convertTableMetadataToColumns(tableMetadata))
-        .rows(
-            rawlsService
-                .entityQuery(user, dataset.storageSourceId(), tableName, maxRows)
-                .getResults()
-                .stream()
-                .map(entity -> convertEntityToRow(entity, tableMetadata.getIdName()))
-                .toList());
-  }
-
-  private Object convertEntityToRow(Entity entity, String idName) {
-    Map<String, Object> att = entity.getAttributes();
-    Map<String, Object> rows = new HashMap<>(att);
-    rows.put(idName, entity.getName());
-    return rows;
-  }
-
-  private static List<ColumnModel> convertTableMetadataToColumns(EntityTypeMetadata entity) {
-    List<ColumnModel> columns = new ArrayList<>();
-    columns.add(new ColumnModel().name(entity.getIdName()));
-    entity.getAttributeNames().stream()
-        .map(name -> new ColumnModel().name(name))
-        .forEach(columns::add);
-    return columns;
   }
 
   private void ensureActionPermission(
@@ -275,31 +195,6 @@ public class DatasetService {
 
   private void validateMetadata(String metadata) {
     toJsonNode(metadata);
-  }
-
-  private static List<TableMetadata> convertDatarepoTablesToCatalogTables(
-      List<TableModel> datarepoTables) {
-    return datarepoTables.stream()
-        .map(
-            tableModel ->
-                new TableMetadata()
-                    .name(tableModel.getName())
-                    .hasData(tableModel.getRowCount() > 0))
-        .toList();
-  }
-
-  private static List<TableMetadata> convertRawlsTablesToCatalogTables(
-      Map<String, EntityTypeMetadata> entityTables) {
-    return entityTables.entrySet().stream()
-        .map(
-            entry ->
-                new TableMetadata().name(entry.getKey()).hasData(entry.getValue().getCount() != 0))
-        .toList();
-  }
-
-  private static ColumnModel convertDatarepoColumnModelToCatalogColumnModel(
-      bio.terra.datarepo.model.ColumnModel datarepoColumnModel) {
-    return new ColumnModel().name(datarepoColumnModel.getName());
   }
 
   public DatasetPreviewTable getDatasetPreview(
